+++ backend/server.ts
@@ -1,4 +1,4 @@
-import app from "./index";
+import app from "./routes/index";
 require("dotenv").config();
 const { MongoClient } = require("mongodb");
